requirements.txt
flask
pillow
flask_wtf
wtforms